pyproject.toml
[tool.poetry]
name = "baratito_backend"
version = "0.1.0"
description = ""
authors = ["app.baratito@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68"
aiofiles = "^0.7.0"
structlog = "^21.1.0"
alembic = "^1.6.5"
SQLAlchemy = "^1.4.22"
psycopg2-binary = "^2.9.1"
dependency-injector = "^4.35.2"
oauth2client = "^4.1.3"
google-auth = "^2.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
GeoAlchemy2 = "^0.9.4"
googlemaps = "^4.5.3"

[tool.poetry.dev-dependencies]
isort = "^5.9.1"
black = "^21.6b0"
flake8 = "^3.9.2"
pytest = "^6.2.4"
cookiecutter = "^1.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.isort]
src_paths = "app"
